fdo#45779 avoiding creation of inconsistent B2DPolygon
this avoid the root cause of this bug, avoiding creating a
B2DPlygon which contains no points.

It seems the code relies somehow on an null B2DPolyPolygon, hence the
change done here. Better would be to have time to look how to remove
this fact. But currently it seems the code relies on  a
rSet.Put( XLineStartItem( aArrowName, aPolyPoly) where aPolyPoly is
not defined in certain cases.

Change-Id: I61b75d925090d1c9a0da96ce1a6eea50a2d60e5a