make CheckCharacterBounds tests happy without forcing empty glyphs 1 unit wide
related, i#87757 and 78cb5f8e318798169fb1ea50c961decf6e06e4f4
which itself is related to e0eced9fbffae86a866b4d41ae7670043c6530b5,

These character bounds are backed by what's the glyph bounding box for the ink
used for the glyph. Whitespace has no ink so its an empty Rectangle. Which
brings the awesome RECT_EMPTY into play.

It might be a bit dubious in the first place to back getCharacterBounds with
the glyph bounding box in the first place, rather than maybe the advance width
or some such. But lets assume that decision was intentional.

So, the qa test should accept that a glyph might be of 0 width anyway.
Then, tweak rectangle merging so that we can preserve the correct top-left
position of an empty glyph
So, we can determine the correct character index given the top-left position
of an empty glyph

Change-Id: I5e18460ff7cd90cd27d5eede2aa0a5494c36a5d3