Replace SvStringsISortDtor in edtwin.cxx and gloslst.[ch]xx
Note that the replacement vector stores all strings first and then
sort and "unique" are applied (ASCII treated as case-insensitive).
Previously strings were inserted sequentially and only the first
version of a string would be stored (case-insensitive ASCII
comparision). This should have no material impact as the strings
retreived from SwAutoCompleteWord are already unique (case-
insensitive ASCII comparison) and the capitalization of the string
is generally changed anyway to match the capitalization of the word
to be auto-completed. Also, there appears to be no logical reason
to store the first inserted version of a string over of the first
version post-sort.

Change-Id: I132865bbb9b382d417fb2cff9de351fdb2cbfb13