fdo#46378: Lanczos3 resampling of images added to Bitmap.
Current resampling methods for images are FAST and INTERPOLATE.
FAST is used where speed of resampling is required, on the other
hand INTERPOLATE resampling is used when we need quality. For
example INTERPOLATE resampling method is used at PDF export.
INTERPOLATE resampling uses bilinear interpolation which is known
to be lower quality as other modern (and slower) resampling
algorithms such as Lanczos, Mitchell or BiCubic resampling.
This change adds Lanczos resampling to the Bitmap class and
enables Lanczos resampling in PDF export.

Lanczos3 resampling is implmented using separable convolution
with which it is also possible to easily add other resampling
methods like BiCubic just by changing the kernel function.

Change-Id: I8dff5b65753b09dffd5bc34f2343d9818efb3e58