bnc#524250: Show backing component after cancelled recovery
...instead of an empty Writer document, which is non-intuitive when the user had
worked on something other than a Writer document before the crash.

64e9649cfb859eaa04f1b272b7c0ef4c4d75f37b "win32-dont-attempt-restart.diff: Don't
attempt to restart OOo after crash" had previously attempted to fix this with a
sledgehammer approach of not restarting at all after a crash on Windows, under
the wrong assumption that there would be no restart after crash at all on Linux
either (which was apparently rather a bug caused by the newly introduced
oosplash at that time).

Change-Id: If65d2e21640d56262c486d887c4a29137cabcd0a