fdo#35365 Set the default font color to black and the document color to white.
The font and document color of a Writer document or an Impress presentation
should not be derived from a desktop theme. A Writer documents needs to look
good on paper. An Impress presentation may have it's own theme. The appearance
of a document should not change by changing the desktop theme.

With this change a document looks the same on any computer if the user do not
change the default colors.

Change-Id: Ia42ca7882f0d2dd1f2a304db5e4b5aaba23244fc