attempt to anchor vml shapes properly as they should be
This is a mess. Shapes from oox/vml are added to Writer by calling
SwXDrawPage::add(), which forces the shapes to be AT_PAGE, because
the pDoc->GetCurrentLayout() there is NULL. After that writerfilter
in DomainMapper_Impl::PushShapeContext() tries to do further
tricks with the anchoring. Such as the fix for bnc#764005
(1bfcc6d14966c2fef58286821876f4b5a77e8daa) that broke inline shapes
because of the AT_PAGE above.

So try to set the proper anchoring already in oox, although I'm not
entirely sure I got it right. Ideally writerfilter should not touch
it, but when I simply removed it another testcase breaks, so hopefully
this all can be cleaned up somewhen later when we have enough testcases
for all the possibilities.

Change-Id: I7dd95d87bd027ce0f129f76b3377f3d1445c34d1