fdo#38116, fdo#49438: fix double hairline border drawing some more:
These two fixes interact in a bad way:
d58467b5a33d700129b1e9166f917a1979e77d6b (fdo#48647)
4eeebee1212e2d9a2a04526a0f9591d916e20261 (fdo#49438)

The problem is that now hairline borders are detected properly, and the
double hairline borders are clipped to the right area, but the clipping
polygon actually collapses to a single line for the hairline border,
with the result that the clipped border line is an empty PolyPolygon
for e.g. a double border 3 twips wide.

Fix this by enforcing a minimum clipping polygon width of 2 units,
which seems to work nicely.

Change-Id: If9cf9f0ed6a5c837860c2313a84cd26b07734b03