Separable Gaussian Blur and Unsharpen Mask filter.
Currently implemented Gaussian Blur filter uses a static matrix
which means that the filter can not be parameterized. The new
implementation of Gaussian Blur filter and accepts a blur radius
as a parameter so the user can change the strength of the blur.
For this the blur matrix is generated at each call. The new Blur
implementation reuses separable convolution from Lanzcos rescale.

For negative values of radius the Bitmap will be sharpened. For
this an Unsharpen Mask filter is used, which is actually a blurred
image substracted from the original image.