ByteBufferWrapper tweaks
operator() gets called in a Java GC thread. A JNIEnv* saved in the
constructor (which is called from the UI thread) is not valid in other
threads (although for now, Dalvik notices, warns, and works around
it). So don't bother keeping the JNIEnv*. Instead fetch one when
needed. Move the method implementations from inline in the header to
libotouch's android.cxx.

Change-Id: I7de6fc54bb8d9a59146576d6e8c325fe917393ee