Added support for the repeatCount attribute for time containers.
The support is limited to the case when the value of the repeatCount attribute is an integral 
number and the duration time is defined implicitly by the active duration of the time container 
children. The accumulate attribute is not handled.

The support for such a feature has been implemented for both the C++ and the JavaScript 
presentation engine.