adapt callers of getCurrFormatsImpl() to changes
4260908f03e3bbcec8cf31cad046187c8e369285 changed String to OUString and
initialized some variables with -1 instead of STRING_NOTFOUND, which was
unsigned 0xffff so var1<var2 was true if var1 had a value other than
default and var2 was default. After the change the comparison had logic
reversed and the result was a default negative currency format with
parentheses for all locales.

Change-Id: I3fe49fed7c1047618dc653fce992eb01582c398d