Show backing window instead of empty writer upon an "empty session"
...(i.e., after logging out of a session with soffice still running, but only
with the backing window still open).  The key issue was that bSessionDataExists
would always be true if /org.openoffice.Office.Recovery/RecoveryInfo/SessionData
is true, even if /org.openoffice.Office.Recovery/RecoveryList is empty.  Always
calling doRestore (unless bExistsRecoveryData, but even if !bExistsSessionData)
is there to ensure that /org.openoffice.Office.Recovery/RecoveryInfo/SessionData
is eventually reset (whether or not that is really necessary).

Change-Id: I9c3fff23f8019069f36e1dd5ded634ea1917bbd4