specify the sourced file with path
"man source" says:

"If filename does not contain a slash, file names  in  PATH  are used to
find the directory containing file‐ name.  The file searched for in PATH
need not  be  executable. When  bash  is  not  in posix  mode,  the
current directory is searched if no file is found in PATH."

Change-Id: I7dc15d30ba925c71d02da2cd8b91ea08192730fc