fdo#46338 preserve sheet name input of invalid reference
If sheet name doesn't exist the resulting reference had replaced the
sheet name with #REF!, e.g. #REF!.A1, because the actual sheet name is
not part of the reference. From a syntactically valid but otherwise
invalid non-external reference create an ocBad token instead to preserve
the sheet name.

Currently a #NAME? error instead of #REF! is generated, further work
would be needed to pass specific error values with ocBad.

Change-Id: I5a608a74d3b3ff2baa4967f2b2e3078cfecfbabc