fdo#39415: sw: fix collapsing border painting:
In a table with collapsing borders, the horizontal borders  need to be
painted "below" the reference line, and not "centered" on the line like
the vertical borders.
d2e53abe322cce72c7632472a8cfa107b00fb291 was apparently an attempt to
fix this, but unfortunately it adjusted both top and bottom borders
with the top border width, which has horrible results if the top and
bottom borders have different widths.

Change-Id: I7605eda6ef8111612ceb1ded79c0811de9ca7b25