fdo#39415: sw: fix collapsing border painting more:
There is a special case for i#9860, which is for first line in follow
table (i.e. continued on new page) without repeated headlines:
Here the bottom border of the cell is used as the top border in case
the cell does not have a top border; this is fixed now with a bit of
refactoring, so the correct half width adjustment can be made.

Change-Id: I58ef81b7e13544014ef75973ea793f673a242488