do not attempt to compile running formula cell (fdo#51601 related)
eb09c0a278aa164efa2efe6674677c4e374bd35b introduced compilation of
a formula cell while it is interpreted, ignoring the comment just above why
that does NOT work ...

This removes the portion that was responsible for the crash. Though import
changed and this did not crash anymore on opening the testcase documents, the
questionable code is still wrong and therefor removed.