Revert "Remove unused throbber.src"
This reverts commit a60a6e21a0a60b6434311f5ad2f4c8dde50b7e5c.

This file isn't referenced in any other file. For me those icon
files are not in any images*.zip file. But the update extension
seems to use the Throbber class, which refers to this icons.

To be absolutely sure I revert the change. Maybe someone else wants
to have a look at this.