tubes: send a file for collaboration when buddy session starts
Channels for file and for tube are independent in telepathy,
so let sender create UUID and pass it to receiver, who then can
bind the document to the channel.
UUID for tube channel goes through telepathy.
UUID for file channel is encoded in the filename for now.

Tubes specific CreateDocFunc is re-introduced, so we could set
current UUID after file is received and when the document is being
constructed, get channel from TeleManager with this UUID.
This is not immune to constructing other documents in the middle of binding
proccess.

Change-Id: I57c7e57a5d7d3ccd7d94677a8cf2719c78baa2fd